requirements.txt
config
mock==3.0.5
mongoengine==0.19.1
codecov
pytest==4.6.9
pytest-cov
pytest-mock==2.0.0
python-dotenv==0.10.5
keyring==18.0.1
coverage
setuptools~=40.8.0
pymongo~=3.10.1
jsonpickle
apscheduler
cryptomodel-athanikos==2.6.1
werkzeug
keyrings.alt
